Update aid type to long and fix abid conversion

+++ v2_spring_backend/tdd/src/main/java/com/bunnyxt/tdd/controller/VideoLogRestController.java
@@ -22,7 +22,7 @@ public class VideoLogRestController {
 
     @RequestMapping(value = "/video/log", method = RequestMethod.GET)
     public ResponseEntity<List<VideoLog>> queryVideoLogs(
-            @RequestParam(defaultValue = "0") Integer aid,
+            @RequestParam(defaultValue = "0") Long aid,
             @RequestParam(defaultValue = "0") Integer start_ts,
             @RequestParam(defaultValue = "0") Integer end_ts,
             @RequestParam(defaultValue = "") String attr,
