+++ v2_spring_backend/tdd/src/main/java/com/bunnyxt/tdd/controller/VideoRestController.java
@@ -23,7 +23,7 @@ public class VideoRestController {
     @PreAuthorize("hasRole('admin')")
     @RequestMapping(value = "/video/{aid}", method = RequestMethod.PUT)
     public Video updateVideoByAid(
-            @PathVariable Integer aid,
+            @PathVariable Long aid,
             @RequestBody Video video
     ) throws InvalidRequestParameterException, IllegalAccessException {
         // when use PUT method, you could provide multiple parameters of video,
