+++ v2_spring_backend/tdd/src/main/java/com/bunnyxt/tdd/dao/TaskDao.java
@@ -2,5 +2,5 @@ package com.bunnyxt.tdd.dao;
 
 public interface TaskDao {
 
-    void addVisitVideoRecord(Integer added, Integer aid, Long userid);
+    void addVisitVideoRecord(Integer added, Long aid, Long userid);
 }
